refaktorerer pensjonsavtaler og innfører visning av offentlig-tp
