+++ src/components/Pensjonsavtaler/OffentligTjenestePensjon/OffentligTjenestepensjon.module.scss.d.ts
@@ -1,0 +1,13 @@
+import globalClassNames from "../../../style.d";
+declare const classNames: typeof globalClassNames & {
+  readonly info: "info";
+  readonly infoIcon: "infoIcon";
+  readonly infoIcon__orange: "infoIcon__orange";
+  readonly infoIcon__blue: "infoIcon__blue";
+  readonly infoText: "infoText";
+  readonly mobileTable: "mobileTable";
+  readonly desktopTableHeader: "desktopTableHeader";
+  readonly desktopTableRader__noBottomBorder: "desktopTableRader__noBottomBorder";
+  readonly desktopTableRader__alignTop: "desktopTableRader__alignTop";
+};
+export = classNames;
