+++ src/components/Pensjonsavtaler/Pensjonsavtaler.module.scss
@@ -1,27 +1,5 @@
 @import '../../scss/variables.scss';
 
-.info {
-  display: flex;
-
-  &Icon {
-    flex-basis: 1.5rem;
-    vertical-align: top;
-    margin-right: var(--a-spacing-2);
-
-    &__orange {
-      color: var(--a-orange-600);
-    }
-
-    &__blue {
-      color: var(--a-deepblue-600);
-    }
-  }
-
-  &Text {
-    flex-basis: 100%;
-  }
-}
-
 .section {
   display: flex;
   flex-direction: column;
@@ -48,4 +26,8 @@
 
 .paragraph {
   margin-bottom: var(--a-spacing-4);
+}
+
+.footnote {
+  padding-top: var(--a-spacing-4);
 }
