+++ src/components/Pensjonsavtaler/Pensjonsavtaler.module.scss.d.ts
@@ -1,14 +1,10 @@
 import globalClassNames from "../../style.d";
 declare const classNames: typeof globalClassNames & {
-  readonly info: "info";
-  readonly infoIcon: "infoIcon";
-  readonly infoIcon__orange: "infoIcon__orange";
-  readonly infoIcon__blue: "infoIcon__blue";
-  readonly infoText: "infoText";
   readonly section: "section";
   readonly "navds-accordion__item": "navds-accordion__item";
   readonly "navds-accordion__item--open": "navds-accordion__item--open";
   readonly sectionHeader: "sectionHeader";
   readonly paragraph: "paragraph";
+  readonly footnote: "footnote";
 };
 export = classNames;
