+++ src/components/Pensjonsavtaler/PrivatePensjonsavtaler/PrivatePensjonsavtalerDesktop.module.scss.d.ts
@@ -1,4 +1,4 @@
-import globalClassNames from "../../style.d";
+import globalClassNames from "../../../style.d";
 declare const classNames: typeof globalClassNames & {
   readonly TableHeader: "TableHeader";
   readonly TableRader__noBottomBorder: "TableRader__noBottomBorder";
