+++ src/components/Pensjonsavtaler/PrivatePensjonsavtaler/PrivatePensjonsavtalerMobile.module.scss
@@ -1,0 +1,6 @@
+@import '../../../scss/variables.scss';
+
+.table {
+  width: 100%;
+  border-spacing: 0;
+}
