+++ src/components/Pensjonsavtaler/PrivatePensjonsavtaler/PrivatePensjonsavtalerMobile.module.scss.d.ts
@@ -1,0 +1,5 @@
+import globalClassNames from "../../../style.d";
+declare const classNames: typeof globalClassNames & {
+  readonly table: "table";
+};
+export = classNames;
