+++ src/components/Pensjonsavtaler/PrivatePensjonsavtaler/index.tsx
@@ -1,0 +1,3 @@
+export { PrivatePensjonsavtalerDesktop } from './PrivatePensjonsavtalerDesktop'
+export { PrivatePensjonsavtalerMobile } from './PrivatePensjonsavtalerMobile'
+export { PrivatePensjonsavtaler } from './PrivatePensjonsavtaler'
