+++ src/components/Pensjonsavtaler/__tests__/Pensjonsavtaler-utils.test.ts
@@ -1,7 +1,5 @@
-import { vi } from 'vitest'
-
 import pensjonsavtalerData from '../../../mocks/data/pensjonsavtaler/67.json' with { type: 'json' }
-import { groupPensjonsavtalerByType, getMaanedString } from '../utils'
+import { groupPensjonsavtalerByType } from '../utils'
 
 describe('GrunnlagPensjonsavtaler-utils', () => {
   const avtalerWithKeys = pensjonsavtalerData.avtaler.map(
@@ -33,26 +31,6 @@ describe('GrunnlagPensjonsavtaler-utils', () => {
       expect(grouped['andre avtaler']).toHaveLength(1)
       expect(grouped['individuelle ordninger']).toHaveLength(2)
       expect(grouped['privat tjenestepensjon']).toHaveLength(3)
-    })
-  })
-
-  describe('getMaanedString', () => {
-    it('returnerer tom streng når måned er undefined eller lik 0', () => {
-      const mockFn = vi.fn()
-      expect(getMaanedString(mockFn)).toEqual('')
-      expect(getMaanedString(mockFn, 0)).toEqual('')
-      expect(mockFn).not.toHaveBeenCalled()
-    })
-    it('returnerer riktig streng når måned er større enn 0', () => {
-      const mockFn = vi.fn().mockReturnValue('string')
-      expect(getMaanedString(mockFn, 1)).toEqual(' string 1 string')
-      expect(mockFn).toHaveBeenNthCalledWith(1, {
-        id: 'string.og',
-      })
-      expect(mockFn).toHaveBeenNthCalledWith(2, {
-        id: 'alder.md',
-      })
-      expect(getMaanedString(mockFn, 5)).toEqual(' string 5 string')
     })
   })
 })
