+++ src/scss/global.scss
@@ -1,7 +1,3 @@
 .nowrap {
   white-space: nowrap;
 }
-
-.full-width {
-  width: 100%;
-}
